+++ package.json
@@ -50,11 +50,11 @@
     "helmet": "^3.21.0",
     "jsonwebtoken": "^8.5.1",
     "jwks-rsa": "^1.6.0",
-    "knex": "^0.19.5",
+    "knex": "^0.21.1",
     "knex-cleaner": "^1.3.0",
     "moment": "^2.24.0",
     "morgan": "^1.9.1",
-    "pg": "^7.12.1",
+    "pg": "^8.1.0",
     "uuid": "^3.3.3"
   },
   "devDependencies": {
