+++ schema/index.js
@@ -1,0 +1,13 @@
+const { gql } = require('apollo-server-express');
+const queryTypes = require('./Query');
+const mutationTypes = require('./Mutation');
+const tableTypes = require('./Tables');
+const inputTypes = require('./Inputs');
+
+const typeDefs = gql`
+  ${queryTypes}
+  ${mutationTypes}
+  ${tableTypes}
+  ${inputTypes}
+`;
+module.exports = typeDefs;
